Code and test data for normalizing labels
There are a number of labels that need to all be normalized to Y, and a bunch of others that all need to be normalized to N. Then there are some (segmentation errors, and ones that I wasn't sure about, for instance) that should not be included in training/testing sets at all. This script takes care of all of that.