Now handles inclusion/exclusion of ones that Mike would not approve of
There are a bunch of instances in the gene-disease data that Mike would not approve of, but that Schank would. The script now has a variable that you can change depending on whether or not you want to include those as Y; if not, they come out as N when normalizing the class labels. Also updated the test input and gold output files.